🐎🔨 large refactor of aggregator for better performance
When thread > grids, we do the mutex locks at the C++/aggregator level
instead of the task part level, this means when we have multiple
aggregators, we can use the CPU more effectively.

+++ packages/vaex-core/src/hash_string.hpp
@@ -378,7 +378,7 @@ class ordered_set : public hash_base<ordered_set<T>, T, T, V> {
                 }
             }
         }
-        if (set->count() != keys->length) {
+        if (set->count() != (int64_t)keys->length) {
             throw std::runtime_error(std::string("key array of length ") + std::to_string(keys->length) + " does not match expected length of " + std::to_string(set->count()));
         }
         if (nan_count == 0) {
